requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
asyncpg>=0.24.0
aioredis>=2.0.0
aiohttp>=3.8.0
python-dateutil>=2.8.2
pytz>=2021.3
black>=21.12b0
isort>=5.10.1
pytest>=6.2.5
pytest-asyncio>=0.16.0
pytest-cov>=3.0.0
aiosqlite>=0.17.0
requests>=2.26.0
Pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.21.0 